==> 1_code/6_Final_Code/5_PCA_PERMANOVA.R <==
library("readr")
library("dplyr")
library("ggplot2"); theme_set(theme_bw() +
                                theme(axis.line = element_line(color='black'),
                                      plot.background = element_blank(),
                                      panel.grid.minor = element_blank(),
                                      panel.grid.major = element_blank()))
library("scales")
library("performance")
library("splines")
library("effects")
library("mgcv")

### To Do: 

### 1: Perform tests on PCA to see if different compartments are truly different.

library("corrplot")
library("vegan")
library("RVAideMemoire")

alldata <- data.frame(read_csv("2_incremental/20220420_STANDING_CROP.csv"))

alldata$SAMPLING_DATE<-as.Date(alldata$SAMPLING_DATE, format = "%m/%d/%Y")

names(alldata)

COMPARTMENTS <- alldata[which(alldata$SAMPLE_DESCRIPTOR == "EPIL"| alldata$SAMPLE_DESCRIPTOR == "EPIP" | alldata$SAMPLE_DESCRIPTOR == "FILA"),]

nrow(COMPARTMENTS)
names(COMPARTMENTS)

####Big 5+Fe+Se+Mo####

LABELS<-as.data.frame(na.omit(COMPARTMENTS[,c(3,4,6,12,13,18,19,20,21,22,25,29,30,31,33,35,36,40,49)]))
LABELS<-LABELS[order(LABELS$SAMPLE_DESCRIPTOR, decreasing = F), ]
EPIL.COR.DF3<-as.data.frame(na.omit(COMPARTMENTS[,c(12,13,18,19,20,21,22,25,29,30,31,33,35,36,40,49)]))
PCA.DF<-as.data.frame(LABELS[,c(4:ncol(LABELS))])
names(PCA.DF)<- c("Al","As","Ca","Cd","Co","Cr","Cu","Fe","Mg","Mn","Mo","Ni","Pb","S","Se","Zn")

#dev.new()
cor(PCA.DF)
corrplot(cor(PCA.DF), method = "circle", lwd=2) 


###PCA###

PCA.DF<-log(PCA.DF)
my.data <- PCA.DF
my.data <- as.matrix(scale(my.data, center = TRUE, scale = TRUE))
groups <- factor(LABELS$SAMPLE_DESCRIPTOR)

my.prc <- prcomp(na.omit(my.data))

plot(my.prc)

dis <- vegdist(my.data, method="euclidean")

pairwise.permanova <- pairwise.perm.manova(dis, groups, nperm = 999, p.method = "fdr", F = FALSE, R2 = FALSE)

mod <- betadisper(dis, groups, sqrt.dist=T, bias.adjust=T, type = "centroid")

pairwise.variance <- permutest(mod, pairwise = TRUE, permutations = 999)

pca_scores<-scores(my.prc)

sd <- my.prc$sdev
correlations <- t(t(my.prc$rotation)*sd)

summary(my.prc)

#####PLOT####


colors <- c(colors()[89], "gold", "chartreuse")
colors <- colors[as.numeric(factor(na.omit(LABELS$SAMPLE_DESCRIPTOR)))]

dev.new()

my.prc$x[,1]<-my.prc$x[,1]*-1
correlations[,1]<-correlations[,1]*-1

plot	(my.prc$x[,1], my.prc$x[,2],
      pch=1,
      xlim=c(-max(abs(range(my.prc$x[,1]))),max(abs(range(my.prc$x[,1])))),	
      ylim=c(-max(abs(range(my.prc$x[,2]))),max(abs(range(my.prc$x[,2])))), 
      cex.lab=1.5, cex.axis=1.5, cex.main=1.4, cex.sub=1.4, 
      col="white",
      xlab="PC 1",
      ylab="PC 2"
)
#	col="black")

tab <- matrix(c(my.prc$x[,1], my.prc$x[,2]), ncol=2)

legend(x = "topleft", legend = levels(factor(LABELS$SAMPLE_DESCRIPTOR)), 
       pch=c(23,24,25), 
       cex = 1, 
       box.lwd=3,
       col = "black", 
       pt.bg=c(colors()[89], "gold", "chartreuse"), pt.lwd=3)


panel.first= {
  ordiellipse(tab,LABELS$SAMPLE_DESCRIPTOR,conf=0.95, draw ="polygon", col=c(adjustcolor(colors()[89], alpha.f=0.25),adjustcolor("gold", alpha.f=0.25),adjustcolor("chartreuse", alpha.f=0.25)), border=c(colors()[89],"gold","chartreuse"), lwd=3)
}

points(my.prc$x[,1],my.prc$x[,2], pch=c(23,24,25)[as.numeric(factor(LABELS$SAMPLE_DESCRIPTOR))],col="black", bg=colors,    cex=1.5,     lwd=2)

par(new=TRUE)

plot	(correlations[1,],
      correlations[2,],
      col="NA",
      xlim=c(-1,1),
      ylim=c(-1,1),
      xaxt="n",
      yaxt="n",
      xlab="",
      ylab="")

arrows(0, 0, 
       x1 = correlations[,1], 
       y1 = correlations[,2], 
       length = 0.1, angle = 30, lwd =2.0,
       code = 2, col = par("fg"), 
       lty = par("lty"))

text(	x=correlations[,1]*1.075,
      y=correlations[,2]*1.075,
      labels = names(PCA.DF), 
      cex=1.0,
      font=2)

rownames(my.prc$rotation)

box(lwd=3)

#legend('topright', legend = levels(wa$Stream), col = 1:4, cex = 1, pch = 19)

coordnames<-data.frame(correlations[,1]*1.1,correlations[,2]*1.1)

coordnames

axis(3, cex.axis=1.5)
mtext("y2",side=4,line=2, col="Blue")
axis(4, cex.axis=1.5)
mtext("y2",side=4,line=2, col="Blue")

abline(h=c(0,0))
abline(v=c(0,0))
